setting up signup service, mappers, exceptions, and account controller

+++ FruitWalletBackend/src/main/java/io/github/larrythexu/FruitWalletBackend/dtos/AccountDTO.java
@@ -5,7 +5,7 @@ import java.util.List;
 public record AccountDTO(
     String username,
     String origin,
-    Double appleBalance,
-    Double bananaBalance,
-    Double orangeBalance,
+    float appleBalance,
+    float bananaBalance,
+    float orangeBalance,
     List<FactoryDTO> factoryDTOList) {}
